@@ -8,41 +8,58 @@ import re
 import sys
 import traceback
 import zipfile
+from collections.abc import AsyncIterator, Awaitable, Callable
+from contextlib import asynccontextmanager
 from functools import lru_cache
 from io import BytesIO, TextIOWrapper
 from pathlib import Path
 from tempfile import NamedTemporaryFile, TemporaryFile
-from typing import Dict, List, Optional
+from typing import Annotated, Any, Literal, Optional
 
 import soundfile
 import uvicorn
-from fastapi import FastAPI, Form, HTTPException, Query, Request, Response
+from fastapi import Body, Depends, FastAPI, Form, HTTPException
+from fastapi import Path as FAPath
+from fastapi import Query, Request, Response
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import HTMLResponse, JSONResponse
+from fastapi.openapi.utils import get_openapi
+from fastapi.responses import JSONResponse
 from fastapi.templating import Jinja2Templates
-from pydantic import ValidationError, conint
+from pydantic import ValidationError, parse_obj_as
 from starlette.background import BackgroundTask
+from starlette.middleware.errors import ServerErrorMiddleware
 from starlette.responses import FileResponse
 
 from voicevox_engine import __version__
 from voicevox_engine.cancellable_engine import CancellableEngine
-from voicevox_engine.downloadable_library import LibraryManager
-from voicevox_engine.engine_manifest import EngineManifestLoader
+from voicevox_engine.core.core_adapter import CoreAdapter
+from voicevox_engine.core.core_initializer import initialize_cores
 from voicevox_engine.engine_manifest.EngineManifest import EngineManifest
-from voicevox_engine.kana_parser import create_kana, parse_kana
-from voicevox_engine.metas.MetasStore import MetasStore, construct_lookup
+from voicevox_engine.engine_manifest.EngineManifestLoader import EngineManifestLoader
+from voicevox_engine.library_manager import LibraryManager
+from voicevox_engine.metas.Metas import StyleId
+from voicevox_engine.metas.MetasStore import (
+    MetasStore,
+    construct_lookup,
+    filter_speakers_and_styles,
+)
 from voicevox_engine.model import (
     AccentPhrase,
     AudioQuery,
-    DownloadableLibrary,
+    BaseLibraryInfo,
+    DownloadableLibraryInfo,
+    FrameAudioQuery,
+    InstalledLibraryInfo,
     MorphableTargetInfo,
     ParseKanaBadRequest,
     ParseKanaError,
+    Score,
     Speaker,
     SpeakerInfo,
-    SpeakerNotFoundError,
+    StyleIdNotFoundError,
     SupportedDevicesInfo,
     UserDictWord,
+    VvlibManifest,
     WordTypes,
 )
 from voicevox_engine.morphing import (
@@ -53,16 +70,18 @@ from voicevox_engine.morphing import (
 from voicevox_engine.morphing import (
     synthesis_morphing_parameter as _synthesis_morphing_parameter,
 )
-from voicevox_engine.part_of_speech_data import MAX_PRIORITY, MIN_PRIORITY
-from voicevox_engine.preset import Preset, PresetError, PresetManager
-from voicevox_engine.setting import (
-    USER_SETTING_PATH,
-    CorsPolicyMode,
-    Setting,
-    SettingLoader,
+from voicevox_engine.preset.Preset import Preset
+from voicevox_engine.preset.PresetError import PresetError
+from voicevox_engine.preset.PresetManager import PresetManager
+from voicevox_engine.setting.Setting import CorsPolicyMode, Setting
+from voicevox_engine.setting.SettingLoader import USER_SETTING_PATH, SettingHandler
+from voicevox_engine.tts_pipeline.kana_converter import create_kana, parse_kana
+from voicevox_engine.tts_pipeline.tts_engine import (
+    TTSEngine,
+    make_tts_engines_from_cores,
 )
-from voicevox_engine.synthesis_engine import SynthesisEngineBase, make_synthesis_engines
-from voicevox_engine.user_dict import (
+from voicevox_engine.user_dict.part_of_speech_data import MAX_PRIORITY, MIN_PRIORITY
+from voicevox_engine.user_dict.user_dict import (
     apply_word,
     delete_word,
     import_user_dict,
@@ -70,14 +89,13 @@ from voicevox_engine.user_dict import (
     rewrite_word,
     update_dict,
 )
-from voicevox_engine.utility import (
+from voicevox_engine.utility.connect_base64_waves import (
     ConnectBase64WavesException,
     connect_base64_waves,
-    delete_file,
-    engine_root,
-    get_latest_core_version,
-    get_save_dir,
 )
+from voicevox_engine.utility.core_version_utility import get_latest_core_version
+from voicevox_engine.utility.path_utility import delete_file, engine_root, get_save_dir
+from voicevox_engine.utility.run_utility import decide_boolean_from_env
 
 
 def b64encode_str(s):
@@ -90,43 +108,68 @@ def set_output_log_utf8() -> None:
     """
     # コンソールがない環境だとNone https://docs.python.org/ja/3/library/sys.html#sys.__stdin__
     if sys.stdout is not None:
-        # 必ずしもreconfigure()が実装されているとは限らない
-        try:
+        if isinstance(sys.stdout, TextIOWrapper):
             sys.stdout.reconfigure(encoding="utf-8")
-        except AttributeError:
+        else:
             # バッファを全て出力する
             sys.stdout.flush()
-            sys.stdout = TextIOWrapper(
-                sys.stdout.buffer, encoding="utf-8", errors="backslashreplace"
-            )
+            try:
+                sys.stdout = TextIOWrapper(
+                    sys.stdout.buffer, encoding="utf-8", errors="backslashreplace"
+                )
+            except AttributeError:
+                # stdout.bufferがない場合は無視
+                pass
     if sys.stderr is not None:
-        try:
+        if isinstance(sys.stderr, TextIOWrapper):
             sys.stderr.reconfigure(encoding="utf-8")
-        except AttributeError:
+        else:
             sys.stderr.flush()
-            sys.stderr = TextIOWrapper(
-                sys.stderr.buffer, encoding="utf-8", errors="backslashreplace"
-            )
+            try:
+                sys.stderr = TextIOWrapper(
+                    sys.stderr.buffer, encoding="utf-8", errors="backslashreplace"
+                )
+            except AttributeError:
+                # stderr.bufferがない場合は無視
+                pass
 
 
 def generate_app(
-    synthesis_engines: Dict[str, SynthesisEngineBase],
+    tts_engines: dict[str, TTSEngine],
+    cores: dict[str, CoreAdapter],
     latest_core_version: str,
-    setting_loader: SettingLoader,
+    setting_loader: SettingHandler,
+    preset_manager: PresetManager,
+    cancellable_engine: CancellableEngine | None = None,
     root_dir: Optional[Path] = None,
     cors_policy_mode: CorsPolicyMode = CorsPolicyMode.localapps,
-    allow_origin: Optional[List[str]] = None,
+    allow_origin: Optional[list[str]] = None,
+    disable_mutable_api: bool = False,
 ) -> FastAPI:
     if root_dir is None:
         root_dir = engine_root()
 
-    default_sampling_rate = synthesis_engines[latest_core_version].default_sampling_rate
+    @asynccontextmanager
+    async def lifespan(app: FastAPI) -> AsyncIterator[None]:
+        update_dict()
+        yield
 
     app = FastAPI(
         title="VOICEVOX Engine",
         description="VOICEVOXの音声合成エンジンです。",
         version=__version__,
+        lifespan=lifespan,
     )
+
+    # 未処理の例外が発生するとCORSMiddlewareが適用されない問題に対するワークアラウンド
+    # ref: https://github.com/VOICEVOX/voicevox_engine/issues/91
+    async def global_execution_handler(request: Request, exc: Exception) -> Response:
+        return JSONResponse(
+            status_code=500,
+            content="Internal Server Error",
+        )
+
+    app.add_middleware(ServerErrorMiddleware, handler=global_execution_handler)
 
     # CORS用のヘッダを生成するミドルウェア
     localhost_regex = "^https?://(localhost|127\\.0\\.0\\.1)(:[0-9]+)?$"
@@ -154,7 +197,9 @@ def generate_app(
 
     # 許可されていないOriginを遮断するミドルウェア
     @app.middleware("http")
-    async def block_origin_middleware(request: Request, call_next):
+    async def block_origin_middleware(
+        request: Request, call_next: Callable[[Request], Awaitable[Response]]
+    ) -> Response | JSONResponse:
         isValidOrigin: bool = False
         if "Origin" not in request.headers:  # Originのない純粋なリクエストの場合
             isValidOrigin = True
@@ -174,17 +219,32 @@ def generate_app(
                 status_code=403, content={"detail": "Origin not allowed"}
             )
 
-    preset_manager = PresetManager(
-        preset_path=root_dir / "presets.yaml",
+    # 許可されていないAPIを無効化する
+    def check_disabled_mutable_api() -> None:
+        if disable_mutable_api:
+            raise HTTPException(
+                status_code=403,
+                detail="エンジンの静的なデータを変更するAPIは無効化されています",
+            )
+
+    engine_manifest_data = EngineManifestLoader(
+        engine_root() / "engine_manifest.json", engine_root()
+    ).load_manifest()
+    library_manager = LibraryManager(
+        get_save_dir() / "installed_libraries",
+        engine_manifest_data.supported_vvlib_manifest_version,
+        engine_manifest_data.brand_name,
+        engine_manifest_data.name,
+        engine_manifest_data.uuid,
     )
-    engine_manifest_loader = EngineManifestLoader(
-        root_dir / "engine_manifest.json", root_dir
-    )
-    library_manager = LibraryManager(get_save_dir() / "installed_libraries")
 
     metas_store = MetasStore(root_dir / "speaker_info")
 
-    setting_ui_template = Jinja2Templates(directory=engine_root() / "ui_template")
+    setting_ui_template = Jinja2Templates(
+        directory=engine_root() / "ui_template",
+        variable_start_string="<JINJA_PRE>",
+        variable_end_string="<JINJA_POST>",
+    )
 
     # キャッシュを有効化
     # モジュール側でlru_cacheを指定するとキャッシュを制御しにくいため、HTTPサーバ側で指定する
@@ -193,19 +253,23 @@ def generate_app(
 
     # @app.on_event("startup")
     # async def start_catch_disconnection():
-    #     if args.enable_cancellable_synthesis:
+    #     if cancellable_engine is not None:
     #         loop = asyncio.get_event_loop()
     #         _ = loop.create_task(cancellable_engine.catch_disconnection())
 
-    @app.on_event("startup")
-    def apply_user_dict():
-        update_dict()
-
-    def get_engine(core_version: Optional[str]) -> SynthesisEngineBase:
+    def get_engine(core_version: Optional[str]) -> TTSEngine:
         if core_version is None:
-            return synthesis_engines[latest_core_version]
-        if core_version in synthesis_engines:
-            return synthesis_engines[core_version]
+            return tts_engines[latest_core_version]
+        if core_version in tts_engines:
+            return tts_engines[core_version]
+        raise HTTPException(status_code=422, detail="不明なバージョンです")
+
+    def get_core(core_version: Optional[str]) -> CoreAdapter:
+        """指定したバージョンのコアを取得する"""
+        if core_version is None:
+            return cores[latest_core_version]
+        if core_version in cores:
+            return cores[core_version]
         raise HTTPException(status_code=422, detail="不明なバージョンです")
 
     @app.post(
@@ -214,12 +278,17 @@ def generate_app(
         tags=["クエリ作成"],
         summary="音声合成用のクエリを作成する",
     )
-    def audio_query(text: str, speaker: int, core_version: Optional[str] = None):
+    def audio_query(
+        text: str,
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> AudioQuery:
         """
-        クエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
+        音声合成用のクエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
         """
         engine = get_engine(core_version)
-        accent_phrases = engine.create_accent_phrases(text, speaker_id=speaker)
+        core = get_core(core_version)
+        accent_phrases = engine.create_accent_phrases(text, style_id)
         return AudioQuery(
             accent_phrases=accent_phrases,
             speedScale=1,
@@ -228,7 +297,7 @@ def generate_app(
             volumeScale=1,
             prePhonemeLength=0.1,
             postPhonemeLength=0.1,
-            outputSamplingRate=default_sampling_rate,
+            outputSamplingRate=core.default_sampling_rate,
             outputStereo=False,
             kana=create_kana(accent_phrases),
         )
@@ -240,12 +309,15 @@ def generate_app(
         summary="音声合成用のクエリをプリセットを用いて作成する",
     )
     def audio_query_from_preset(
-        text: str, preset_id: int, core_version: Optional[str] = None
-    ):
+        text: str,
+        preset_id: int,
+        core_version: str | None = None,
+    ) -> AudioQuery:
         """
-        クエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
+        音声合成用のクエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
         """
         engine = get_engine(core_version)
+        core = get_core(core_version)
         try:
             presets = preset_manager.load_presets()
         except PresetError as err:
@@ -255,11 +327,11 @@ def generate_app(
                 selected_preset = preset
                 break
         else:
-            raise HTTPException(status_code=422, detail="該当するプリセットIDが見つかりません")
+            raise HTTPException(
+                status_code=422, detail="該当するプリセットIDが見つかりません"
+            )
 
-        accent_phrases = engine.create_accent_phrases(
-            text, speaker_id=selected_preset.style_id
-        )
+        accent_phrases = engine.create_accent_phrases(text, selected_preset.style_id)
         return AudioQuery(
             accent_phrases=accent_phrases,
             speedScale=selected_preset.speedScale,
@@ -268,14 +340,14 @@ def generate_app(
             volumeScale=selected_preset.volumeScale,
             prePhonemeLength=selected_preset.prePhonemeLength,
             postPhonemeLength=selected_preset.postPhonemeLength,
-            outputSamplingRate=default_sampling_rate,
+            outputSamplingRate=core.default_sampling_rate,
             outputStereo=False,
             kana=create_kana(accent_phrases),
         )
 
     @app.post(
         "/accent_phrases",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="テキストからアクセント句を得る",
         responses={
@@ -287,13 +359,13 @@ def generate_app(
     )
     def accent_phrases(
         text: str,
-        speaker: int,
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
         is_kana: bool = False,
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
         """
         テキストからアクセント句を得ます。
-        is_kanaが`true`のとき、テキストは次のようなAquesTalkライクな記法に従う読み仮名として処理されます。デフォルトは`false`です。
+        is_kanaが`true`のとき、テキストは次のAquesTalk 風記法で解釈されます。デフォルトは`false`です。
         * 全てのカナはカタカナで記述される
         * アクセント句は`/`または`、`で区切る。`、`で区切った場合に限り無音区間が挿入される。
         * カナの手前に`_`を入れるとそのカナは無声化される
@@ -303,65 +375,55 @@ def generate_app(
         engine = get_engine(core_version)
         if is_kana:
             try:
-                accent_phrases = parse_kana(text)
+                return engine.create_accent_phrases_from_kana(text, style_id)
             except ParseKanaError as err:
                 raise HTTPException(
-                    status_code=400,
-                    detail=ParseKanaBadRequest(err).dict(),
+                    status_code=400, detail=ParseKanaBadRequest(err).dict()
                 )
-            accent_phrases = engine.replace_mora_data(
-                accent_phrases=accent_phrases, speaker_id=speaker
-            )
-
-            return accent_phrases
         else:
-            return engine.create_accent_phrases(text, speaker_id=speaker)
+            return engine.create_accent_phrases(text, style_id)
 
     @app.post(
         "/mora_data",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="アクセント句から音高・音素長を得る",
     )
     def mora_data(
-        accent_phrases: List[AccentPhrase],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
+        accent_phrases: list[AccentPhrase],
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
         engine = get_engine(core_version)
-        return engine.replace_mora_data(accent_phrases, speaker_id=speaker)
+        return engine.update_length_and_pitch(accent_phrases, style_id)
 
     @app.post(
         "/mora_length",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="アクセント句から音素長を得る",
     )
     def mora_length(
-        accent_phrases: List[AccentPhrase],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
+        accent_phrases: list[AccentPhrase],
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
         engine = get_engine(core_version)
-        return engine.replace_phoneme_length(
-            accent_phrases=accent_phrases, speaker_id=speaker
-        )
+        return engine.update_length(accent_phrases, style_id)
 
     @app.post(
         "/mora_pitch",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="アクセント句から音高を得る",
     )
     def mora_pitch(
-        accent_phrases: List[AccentPhrase],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
+        accent_phrases: list[AccentPhrase],
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
         engine = get_engine(core_version)
-        return engine.replace_mora_pitch(
-            accent_phrases=accent_phrases, speaker_id=speaker
-        )
+        return engine.update_pitch(accent_phrases, style_id)
 
     @app.post(
         "/synthesis",
@@ -378,18 +440,16 @@ def generate_app(
     )
     def synthesis(
         query: AudioQuery,
-        speaker: int,
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
         enable_interrogative_upspeak: bool = Query(  # noqa: B008
             default=True,
             description="疑問系のテキストが与えられたら語尾を自動調整する",
         ),
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> FileResponse:
         engine = get_engine(core_version)
-        wave = engine.synthesis(
-            query=query,
-            speaker_id=speaker,
-            enable_interrogative_upspeak=enable_interrogative_upspeak,
+        wave = engine.synthesize_wave(
+            query, style_id, enable_interrogative_upspeak=enable_interrogative_upspeak
         )
 
         with NamedTemporaryFile(delete=False) as f:
@@ -418,20 +478,17 @@ def generate_app(
     )
     def cancellable_synthesis(
         query: AudioQuery,
-        speaker: int,
         request: Request,
-        core_version: Optional[str] = None,
-    ):
-        if not args.enable_cancellable_synthesis:
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> FileResponse:
+        if cancellable_engine is None:
             raise HTTPException(
                 status_code=404,
                 detail="実験的機能はデフォルトで無効になっています。使用するには引数を指定してください。",
             )
         f_name = cancellable_engine._synthesis_impl(
-            query=query,
-            speaker_id=speaker,
-            request=request,
-            core_version=core_version,
+            query, style_id, request, core_version=core_version
         )
         if f_name == "":
             raise HTTPException(status_code=422, detail="不明なバージョンです")
@@ -458,27 +515,24 @@ def generate_app(
         summary="複数まとめて音声合成する",
     )
     def multi_synthesis(
-        queries: List[AudioQuery],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
+        queries: list[AudioQuery],
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> FileResponse:
         engine = get_engine(core_version)
         sampling_rate = queries[0].outputSamplingRate
 
         with NamedTemporaryFile(delete=False) as f:
-
             with zipfile.ZipFile(f, mode="a") as zip_file:
-
                 for i in range(len(queries)):
-
                     if queries[i].outputSamplingRate != sampling_rate:
                         raise HTTPException(
-                            status_code=422, detail="サンプリングレートが異なるクエリがあります"
+                            status_code=422,
+                            detail="サンプリングレートが異なるクエリがあります",
                         )
 
                     with TemporaryFile() as wav_file:
-
-                        wave = engine.synthesis(query=queries[i], speaker_id=speaker)
+                        wave = engine.synthesize_wave(queries[i], style_id)
                         soundfile.write(
                             file=wav_file,
                             data=wave,
@@ -496,35 +550,35 @@ def generate_app(
 
     @app.post(
         "/morphable_targets",
-        response_model=List[Dict[str, MorphableTargetInfo]],
+        response_model=list[dict[str, MorphableTargetInfo]],
         tags=["音声合成"],
-        summary="指定した話者に対してエンジン内の話者がモーフィングが可能か判定する",
+        summary="指定したスタイルに対してエンジン内の話者がモーフィングが可能か判定する",
     )
     def morphable_targets(
-        base_speakers: List[int],
-        core_version: Optional[str] = None,
-    ):
+        base_style_ids: list[StyleId], core_version: str | None = None
+    ) -> list[dict[str, MorphableTargetInfo]]:
         """
-        指定されたベース話者に対してエンジン内の各話者がモーフィング機能を利用可能か返します。
+        指定されたベーススタイルに対してエンジン内の各話者がモーフィング機能を利用可能か返します。
         モーフィングの許可/禁止は`/speakers`の`speaker.supported_features.synthesis_morphing`に記載されています。
         プロパティが存在しない場合は、モーフィングが許可されているとみなします。
-        返り値の話者はstring型なので注意。
+        返り値のスタイルIDはstring型なので注意。
         """
-        engine = get_engine(core_version)
+        core = get_core(core_version)
 
         try:
-            speakers = metas_store.load_combined_metas(engine=engine)
+            speakers = metas_store.load_combined_metas(core=core)
             morphable_targets = get_morphable_targets(
-                speakers=speakers, base_speakers=base_speakers
+                speakers=speakers, base_style_ids=base_style_ids
             )
             # jsonはint型のキーを持てないので、string型に変換する
             return [
                 {str(k): v for k, v in morphable_target.items()}
                 for morphable_target in morphable_targets
             ]
-        except SpeakerNotFoundError as e:
+        except StyleIdNotFoundError as e:
             raise HTTPException(
-                status_code=404, detail=f"該当する話者(speaker={e.speaker})が見つかりません"
+                status_code=404,
+                detail=f"該当するスタイル(style_id={e.style_id})が見つかりません",
             )
 
     @app.post(
@@ -538,43 +592,46 @@ def generate_app(
             }
         },
         tags=["音声合成"],
-        summary="2人の話者でモーフィングした音声を合成する",
+        summary="2種類のスタイルでモーフィングした音声を合成する",
     )
     def _synthesis_morphing(
         query: AudioQuery,
-        base_speaker: int,
-        target_speaker: int,
+        base_style_id: StyleId = Query(alias="base_speaker"),  # noqa: B008
+        target_style_id: StyleId = Query(alias="target_speaker"),  # noqa: B008
         morph_rate: float = Query(..., ge=0.0, le=1.0),  # noqa: B008
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> FileResponse:
         """
-        指定された2人の話者で音声を合成、指定した割合でモーフィングした音声を得ます。
-        モーフィングの割合は`morph_rate`で指定でき、0.0でベースの話者、1.0でターゲットの話者に近づきます。
+        指定された2種類のスタイルで音声を合成、指定した割合でモーフィングした音声を得ます。
+        モーフィングの割合は`morph_rate`で指定でき、0.0でベースのスタイル、1.0でターゲットのスタイルに近づきます。
         """
         engine = get_engine(core_version)
+        core = get_core(core_version)
 
         try:
-            speakers = metas_store.load_combined_metas(engine=engine)
+            speakers = metas_store.load_combined_metas(core=core)
             speaker_lookup = construct_lookup(speakers=speakers)
             is_permitted = is_synthesis_morphing_permitted(
-                speaker_lookup, base_speaker, target_speaker
+                speaker_lookup, base_style_id, target_style_id
             )
             if not is_permitted:
                 raise HTTPException(
                     status_code=400,
-                    detail="指定された話者ペアでのモーフィングはできません",
+                    detail="指定されたスタイルペアでのモーフィングはできません",
                 )
-        except SpeakerNotFoundError as e:
+        except StyleIdNotFoundError as e:
             raise HTTPException(
-                status_code=404, detail=f"該当する話者(speaker={e.speaker})が見つかりません"
+                status_code=404,
+                detail=f"該当するスタイル(style_id={e.style_id})が見つかりません",
             )
 
         # 生成したパラメータはキャッシュされる
         morph_param = synthesis_morphing_parameter(
             engine=engine,
+            core=core,
             query=query,
-            base_speaker=base_speaker,
-            target_speaker=target_speaker,
+            base_style_id=base_style_id,
+            target_style_id=target_style_id,
         )
 
         morph_wave = synthesis_morphing(
@@ -599,6 +656,69 @@ def generate_app(
         )
 
     @app.post(
+        "/sing_frame_audio_query",
+        response_model=FrameAudioQuery,
+        tags=["クエリ作成"],
+        summary="歌唱音声合成用のクエリを作成する",
+    )
+    def sing_frame_audio_query(
+        score: Score,
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> FrameAudioQuery:
+        """
+        歌唱音声合成用のクエリの初期値を得ます。ここで得られたクエリはそのまま歌唱音声合成に利用できます。各値の意味は`Schemas`を参照してください。
+        """
+        engine = get_engine(core_version)
+        core = get_core(core_version)
+        phonemes, f0, volume = engine.create_sing_phoneme_and_f0_and_volume(
+            score, style_id
+        )
+
+        return FrameAudioQuery(
+            f0=f0,
+            volume=volume,
+            phonemes=phonemes,
+            volumeScale=1,
+            outputSamplingRate=core.default_sampling_rate,
+            outputStereo=False,
+        )
+
+    @app.post(
+        "/frame_synthesis",
+        response_class=FileResponse,
+        responses={
+            200: {
+                "content": {
+                    "audio/wav": {"schema": {"type": "string", "format": "binary"}}
+                },
+            }
+        },
+        tags=["音声合成"],
+    )
+    def frame_synthesis(
+        query: FrameAudioQuery,
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> FileResponse:
+        """
+        歌唱音声合成を行います。
+        """
+        engine = get_engine(core_version)
+        wave = engine.frame_synthsize_wave(query, style_id)
+
+        with NamedTemporaryFile(delete=False) as f:
+            soundfile.write(
+                file=f, data=wave, samplerate=query.outputSamplingRate, format="WAV"
+            )
+
+        return FileResponse(
+            f.name,
+            media_type="audio/wav",
+            background=BackgroundTask(delete_file, f.name),
+        )
+
+    @app.post(
         "/connect_waves",
         response_class=FileResponse,
         responses={
@@ -611,14 +731,14 @@ def generate_app(
         tags=["その他"],
         summary="base64エンコードされた複数のwavデータを一つに結合する",
     )
-    def connect_waves(waves: List[str]):
+    def connect_waves(waves: list[str]) -> FileResponse:
         """
         base64エンコードされたwavデータを一纏めにし、wavファイルで返します。
         """
         try:
             waves_nparray, sampling_rate = connect_base64_waves(waves)
         except ConnectBase64WavesException as err:
-            return HTTPException(status_code=422, detail=str(err))
+            raise HTTPException(status_code=422, detail=str(err))
 
         with NamedTemporaryFile(delete=False) as f:
             soundfile.write(
@@ -634,15 +754,15 @@ def generate_app(
             background=BackgroundTask(delete_file, f.name),
         )
 
-    @app.get("/presets", response_model=List[Preset], tags=["その他"])
-    def get_presets():
+    @app.get(
+        "/presets",
+        response_model=list[Preset],
+        response_description="プリセットのリスト",
+        tags=["その他"],
+    )
+    def get_presets() -> list[Preset]:
         """
         エンジンが保持しているプリセットの設定を返します
-
-        Returns
-        -------
-        presets: List[Preset]
-            プリセットのリスト
         """
         try:
             presets = preset_manager.load_presets()
@@ -650,21 +770,23 @@ def generate_app(
             raise HTTPException(status_code=422, detail=str(err))
         return presets
 
-    @app.post("/add_preset", response_model=int, tags=["その他"])
-    def add_preset(preset: Preset):
+    @app.post(
+        "/add_preset",
+        response_model=int,
+        response_description="追加したプリセットのプリセットID",
+        tags=["その他"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def add_preset(
+        preset: Annotated[
+            Preset,
+            Body(
+                description="新しいプリセット。プリセットIDが既存のものと重複している場合は、新規のプリセットIDが採番されます。"
+            ),
+        ]
+    ) -> int:
         """
         新しいプリセットを追加します
-
-        Parameters
-        -------
-        preset: Preset
-            新しいプリセット。
-            プリセットIDが既存のものと重複している場合は、新規のプリセットIDが採番されます。
-
-        Returns
-        -------
-        id: int
-            追加したプリセットのプリセットID
         """
         try:
             id = preset_manager.add_preset(preset)
@@ -672,21 +794,23 @@ def generate_app(
             raise HTTPException(status_code=422, detail=str(err))
         return id
 
-    @app.post("/update_preset", response_model=int, tags=["その他"])
-    def update_preset(preset: Preset):
+    @app.post(
+        "/update_preset",
+        response_model=int,
+        response_description="更新したプリセットのプリセットID",
+        tags=["その他"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def update_preset(
+        preset: Annotated[
+            Preset,
+            Body(
+                description="更新するプリセット。プリセットIDが更新対象と一致している必要があります。"
+            ),
+        ]
+    ) -> int:
         """
         既存のプリセットを更新します
-
-        Parameters
-        -------
-        preset: Preset
-            更新するプリセット。
-            プリセットIDが更新対象と一致している必要があります。
-
-        Returns
-        -------
-        id: int
-            更新したプリセットのプリセットID
         """
         try:
             id = preset_manager.update_preset(preset)
@@ -694,16 +818,17 @@ def generate_app(
             raise HTTPException(status_code=422, detail=str(err))
         return id
 
-    @app.post("/delete_preset", status_code=204, tags=["その他"])
-    def delete_preset(id: int):
+    @app.post(
+        "/delete_preset",
+        status_code=204,
+        tags=["その他"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def delete_preset(
+        id: Annotated[int, Query(description="削除するプリセットのプリセットID")]
+    ) -> Response:
         """
         既存のプリセットを削除します
-
-        Parameters
-        -------
-        id: int
-            削除するプリセットのプリセットID
-
         """
         try:
             preset_manager.delete_preset(id)
@@ -715,68 +840,103 @@ def generate_app(
     def version() -> str:
         return __version__
 
-    @app.get("/core_versions", response_model=List[str], tags=["その他"])
-    def core_versions() -> List[str]:
+    @app.get("/core_versions", response_model=list[str], tags=["その他"])
+    def core_versions() -> Response:
         return Response(
-            content=json.dumps(list(synthesis_engines.keys())),
+            content=json.dumps(list(cores.keys())),
             media_type="application/json",
         )
 
-    @app.get("/speakers", response_model=List[Speaker], tags=["その他"])
+    @app.get("/speakers", response_model=list[Speaker], tags=["その他"])
     def speakers(
-        core_version: Optional[str] = None,
-    ):
-        engine = get_engine(core_version)
-        return metas_store.load_combined_metas(engine=engine)
+        core_version: str | None = None,
+    ) -> list[Speaker]:
+        speakers = metas_store.load_combined_metas(get_core(core_version))
+        return filter_speakers_and_styles(speakers, "speaker")
 
     @app.get("/speaker_info", response_model=SpeakerInfo, tags=["その他"])
-    def speaker_info(speaker_uuid: str, core_version: Optional[str] = None):
+    def speaker_info(
+        speaker_uuid: str,
+        core_version: str | None = None,
+    ) -> SpeakerInfo:
         """
         指定されたspeaker_uuidに関する情報をjson形式で返します。
         画像や音声はbase64エンコードされたものが返されます。
-
-        Returns
-        -------
-        ret_data: SpeakerInfo
         """
-        speakers = json.loads(get_engine(core_version).speakers)
+        return _speaker_info(
+            speaker_uuid=speaker_uuid,
+            speaker_or_singer="speaker",
+            core_version=core_version,
+        )
+
+    # FIXME: この関数をどこかに切り出す
+    def _speaker_info(
+        speaker_uuid: str,
+        speaker_or_singer: Literal["speaker", "singer"],
+        core_version: str | None,
+    ) -> SpeakerInfo:
+        # エンジンに含まれる話者メタ情報は、次のディレクトリ構造に従わなければならない：
+        # {root_dir}/
+        #   speaker_info/
+        #       {speaker_uuid_0}/
+        #           policy.md
+        #           portrait.png
+        #           icons/
+        #               {id_0}.png
+        #               {id_1}.png
+        #               ...
+        #           portraits/
+        #               {id_0}.png
+        #               {id_1}.png
+        #               ...
+        #           voice_samples/
+        #               {id_0}_001.wav
+        #               {id_0}_002.wav
+        #               {id_0}_003.wav
+        #               {id_1}_001.wav
+        #               ...
+        #       {speaker_uuid_1}/
+        #           ...
+
+        # 該当話者の検索
+        speakers = parse_obj_as(
+            list[Speaker], json.loads(get_core(core_version).speakers)
+        )
+        speakers = filter_speakers_and_styles(speakers, speaker_or_singer)
         for i in range(len(speakers)):
-            if speakers[i]["speaker_uuid"] == speaker_uuid:
+            if speakers[i].speaker_uuid == speaker_uuid:
                 speaker = speakers[i]
                 break
         else:
             raise HTTPException(status_code=404, detail="該当する話者が見つかりません")
 
         try:
-            policy = (root_dir / f"speaker_info/{speaker_uuid}/policy.md").read_text(
-                "utf-8"
-            )
-            portrait = b64encode_str(
-                (root_dir / f"speaker_info/{speaker_uuid}/portrait.png").read_bytes()
-            )
+            speaker_path = root_dir / "speaker_info" / speaker_uuid
+            # 話者情報の取得
+            # speaker policy
+            policy_path = speaker_path / "policy.md"
+            policy = policy_path.read_text("utf-8")
+            # speaker portrait
+            portrait_path = speaker_path / "portrait.png"
+            portrait = b64encode_str(portrait_path.read_bytes())
+            # スタイル情報の取得
             style_infos = []
-            for style in speaker["styles"]:
-                id = style["id"]
-                icon = b64encode_str(
-                    (
-                        root_dir / f"speaker_info/{speaker_uuid}/icons/{id}.png"
-                    ).read_bytes()
-                )
-                style_portrait_path = (
-                    root_dir / f"speaker_info/{speaker_uuid}/portraits/{id}.png"
-                )
-                style_portrait = (
-                    b64encode_str(style_portrait_path.read_bytes())
-                    if style_portrait_path.exists()
-                    else None
-                )
+            for style in speaker.styles:
+                id = style.id
+                # style icon
+                style_icon_path = speaker_path / "icons" / f"{id}.png"
+                icon = b64encode_str(style_icon_path.read_bytes())
+                # style portrait
+                style_portrait_path = speaker_path / "portraits" / f"{id}.png"
+                style_portrait = None
+                if style_portrait_path.exists():
+                    style_portrait = b64encode_str(style_portrait_path.read_bytes())
+                # voice samples
                 voice_samples = [
                     b64encode_str(
                         (
-                            root_dir
-                            / "speaker_info/{}/voice_samples/{}_{}.wav".format(
-                                speaker_uuid, id, str(j + 1).zfill(3)
-                            )
+                            speaker_path
+                            / "voice_samples/{}_{}.wav".format(id, str(j + 1).zfill(3))
                         ).read_bytes()
                     )
                     for j in range(3)
@@ -793,138 +953,193 @@ def generate_app(
             import traceback
 
             traceback.print_exc()
-            raise HTTPException(status_code=500, detail="追加情報が見つかりませんでした")
+            raise HTTPException(
+                status_code=500, detail="追加情報が見つかりませんでした"
+            )
 
-        ret_data = {"policy": policy, "portrait": portrait, "style_infos": style_infos}
+        ret_data = SpeakerInfo(
+            policy=policy,
+            portrait=portrait,
+            style_infos=style_infos,
+        )
         return ret_data
 
-    @app.get(
-        "/downloadable_libraries",
-        response_model=List[DownloadableLibrary],
-        tags=["音声ライブラリ管理"],
-    )
-    def downloadable_libraries():
-        """
-        ダウンロード可能な音声ライブラリの情報を返します。
+    @app.get("/singers", response_model=list[Speaker], tags=["その他"])
+    def singers(
+        core_version: str | None = None,
+    ) -> list[Speaker]:
+        singers = metas_store.load_combined_metas(get_core(core_version))
+        return filter_speakers_and_styles(singers, "singer")
 
-        Returns
-        -------
-        ret_data: List[DownloadableLibrary]
+    @app.get("/singer_info", response_model=SpeakerInfo, tags=["その他"])
+    def singer_info(
+        speaker_uuid: str,
+        core_version: str | None = None,
+    ) -> SpeakerInfo:
         """
-        manifest = engine_manifest_loader.load_manifest()
-        if not manifest.supported_features.manage_library:
-            raise HTTPException(status_code=404, detail="この機能は実装されていません")
-        return library_manager.downloadable_libraries()
-
-    @app.get(
-        "/installed_libraries",
-        response_model=List[DownloadableLibrary],
-        tags=["音声ライブラリ管理"],
-    )
-    def installed_libraries():
+        指定されたspeaker_uuidに関する情報をjson形式で返します。
+        画像や音声はbase64エンコードされたものが返されます。
         """
-        インストールした音声ライブラリの情報を返します。
-
-        Returns
-        -------
-        ret_data: List[DownloadableLibrary]
-        """
-        manifest = engine_manifest_loader.load_manifest()
-        if not manifest.supported_features.manage_library:
-            raise HTTPException(status_code=404, detail="この機能は実装されていません")
-        return library_manager.installed_libraries()
-
-    @app.post(
-        "/install_library/{library_uuid}",
-        status_code=204,
-        tags=["音声ライブラリ管理"],
-    )
-    async def install_library(library_uuid: str, request: Request):
-        """
-        音声ライブラリをインストールします。
-        音声ライブラリのZIPファイルをリクエストボディとして送信してください。
-
-        Parameters
-        ----------
-        library_uuid: str
-            音声ライブラリのID
-        """
-        manifest = engine_manifest_loader.load_manifest()
-        if not manifest.supported_features.manage_library:
-            raise HTTPException(status_code=404, detail="この機能は実装されていません")
-        archive = BytesIO(await request.body())
-        loop = asyncio.get_event_loop()
-        await loop.run_in_executor(
-            None, library_manager.install_library, library_uuid, archive
+        return _speaker_info(
+            speaker_uuid=speaker_uuid,
+            speaker_or_singer="singer",
+            core_version=core_version,
         )
-        return Response(status_code=204)
+
+    if engine_manifest_data.supported_features.manage_library:
+
+        @app.get(
+            "/downloadable_libraries",
+            response_model=list[DownloadableLibraryInfo],
+            response_description="ダウンロード可能な音声ライブラリの情報リスト",
+            tags=["音声ライブラリ管理"],
+        )
+        def downloadable_libraries() -> list[DownloadableLibraryInfo]:
+            """
+            ダウンロード可能な音声ライブラリの情報を返します。
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(
+                    status_code=404, detail="この機能は実装されていません"
+                )
+            return library_manager.downloadable_libraries()
+
+        @app.get(
+            "/installed_libraries",
+            response_model=dict[str, InstalledLibraryInfo],
+            response_description="インストールした音声ライブラリの情報",
+            tags=["音声ライブラリ管理"],
+        )
+        def installed_libraries() -> dict[str, InstalledLibraryInfo]:
+            """
+            インストールした音声ライブラリの情報を返します。
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(
+                    status_code=404, detail="この機能は実装されていません"
+                )
+            return library_manager.installed_libraries()
+
+        @app.post(
+            "/install_library/{library_uuid}",
+            status_code=204,
+            tags=["音声ライブラリ管理"],
+            dependencies=[Depends(check_disabled_mutable_api)],
+        )
+        async def install_library(
+            library_uuid: Annotated[str, FAPath(description="音声ライブラリのID")],
+            request: Request,
+        ) -> Response:
+            """
+            音声ライブラリをインストールします。
+            音声ライブラリのZIPファイルをリクエストボディとして送信してください。
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(
+                    status_code=404, detail="この機能は実装されていません"
+                )
+            archive = BytesIO(await request.body())
+            loop = asyncio.get_event_loop()
+            await loop.run_in_executor(
+                None, library_manager.install_library, library_uuid, archive
+            )
+            return Response(status_code=204)
+
+        @app.post(
+            "/uninstall_library/{library_uuid}",
+            status_code=204,
+            tags=["音声ライブラリ管理"],
+            dependencies=[Depends(check_disabled_mutable_api)],
+        )
+        def uninstall_library(
+            library_uuid: Annotated[str, FAPath(description="音声ライブラリのID")]
+        ) -> Response:
+            """
+            音声ライブラリをアンインストールします。
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(
+                    status_code=404, detail="この機能は実装されていません"
+                )
+            library_manager.uninstall_library(library_uuid)
+            return Response(status_code=204)
 
     @app.post("/initialize_speaker", status_code=204, tags=["その他"])
     def initialize_speaker(
-        speaker: int,
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
         skip_reinit: bool = Query(  # noqa: B008
-            False, description="既に初期化済みの話者の再初期化をスキップするかどうか"
+            default=False,
+            description="既に初期化済みのスタイルの再初期化をスキップするかどうか",
         ),
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> Response:
         """
-        指定されたspeaker_idの話者を初期化します。
+        指定されたスタイルを初期化します。
         実行しなくても他のAPIは使用できますが、初回実行時に時間がかかることがあります。
         """
-        engine = get_engine(core_version)
-        engine.initialize_speaker_synthesis(speaker_id=speaker, skip_reinit=skip_reinit)
+        core = get_core(core_version)
+        core.initialize_style_id_synthesis(style_id, skip_reinit=skip_reinit)
         return Response(status_code=204)
 
     @app.get("/is_initialized_speaker", response_model=bool, tags=["その他"])
-    def is_initialized_speaker(speaker: int, core_version: Optional[str] = None):
+    def is_initialized_speaker(
+        style_id: StyleId = Query(alias="speaker"),  # noqa: B008
+        core_version: str | None = None,
+    ) -> bool:
         """
-        指定されたspeaker_idの話者が初期化されているかどうかを返します。
+        指定されたスタイルが初期化されているかどうかを返します。
         """
-        engine = get_engine(core_version)
-        return engine.is_initialized_speaker_synthesis(speaker)
+        core = get_core(core_version)
+        return core.is_initialized_style_id_synthesis(style_id)
 
-    @app.get("/user_dict", response_model=Dict[str, UserDictWord], tags=["ユーザー辞書"])
-    def get_user_dict_words():
+    @app.get(
+        "/user_dict",
+        response_model=dict[str, UserDictWord],
+        response_description="単語のUUIDとその詳細",
+        tags=["ユーザー辞書"],
+    )
+    def get_user_dict_words() -> dict[str, UserDictWord]:
         """
         ユーザー辞書に登録されている単語の一覧を返します。
         単語の表層形(surface)は正規化済みの物を返します。
-
-        Returns
-        -------
-        Dict[str, UserDictWord]
-            単語のUUIDとその詳細
         """
         try:
             return read_dict()
         except Exception:
             traceback.print_exc()
-            raise HTTPException(status_code=422, detail="辞書の読み込みに失敗しました。")
+            raise HTTPException(
+                status_code=422, detail="辞書の読み込みに失敗しました。"
+            )
 
-    @app.post("/user_dict_word", response_model=str, tags=["ユーザー辞書"])
+    @app.post(
+        "/user_dict_word",
+        response_model=str,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def add_user_dict_word(
-        surface: str,
-        pronunciation: str,
-        accent_type: int,
-        word_type: Optional[WordTypes] = None,
-        priority: Optional[conint(ge=MIN_PRIORITY, le=MAX_PRIORITY)] = None,
-    ):
+        surface: Annotated[str, Query(description="言葉の表層形")],
+        pronunciation: Annotated[str, Query(description="言葉の発音（カタカナ）")],
+        accent_type: Annotated[
+            int, Query(description="アクセント型（音が下がる場所を指す）")
+        ],
+        word_type: Annotated[
+            WordTypes | None,
+            Query(
+                description="PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか"
+            ),
+        ] = None,
+        priority: Annotated[
+            int | None,
+            Query(
+                ge=MIN_PRIORITY,
+                le=MAX_PRIORITY,
+                description="単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨",
+            ),
+        ] = None,
+    ) -> Response:
         """
         ユーザー辞書に言葉を追加します。
-
-        Parameters
-        ----------
-        surface : str
-            言葉の表層形
-        pronunciation: str
-            言葉の発音（カタカナ）
-        accent_type: int
-            アクセント型（音が下がる場所を指す）
-        word_type: WordTypes, optional
-            PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか
-        priority: int, optional
-            単語の優先度（0から10までの整数）
-            数字が大きいほど優先度が高くなる
-            1から9までの値を指定することを推奨
         """
         try:
             word_uuid = apply_word(
@@ -936,39 +1151,45 @@ def generate_app(
             )
             return Response(content=word_uuid)
         except ValidationError as e:
-            raise HTTPException(status_code=422, detail="パラメータに誤りがあります。\n" + str(e))
+            raise HTTPException(
+                status_code=422, detail="パラメータに誤りがあります。\n" + str(e)
+            )
         except Exception:
             traceback.print_exc()
-            raise HTTPException(status_code=422, detail="ユーザー辞書への追加に失敗しました。")
+            raise HTTPException(
+                status_code=422, detail="ユーザー辞書への追加に失敗しました。"
+            )
 
-    @app.put("/user_dict_word/{word_uuid}", status_code=204, tags=["ユーザー辞書"])
+    @app.put(
+        "/user_dict_word/{word_uuid}",
+        status_code=204,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def rewrite_user_dict_word(
-        surface: str,
-        pronunciation: str,
-        accent_type: int,
-        word_uuid: str,
-        word_type: Optional[WordTypes] = None,
-        priority: Optional[conint(ge=MIN_PRIORITY, le=MAX_PRIORITY)] = None,
-    ):
+        surface: Annotated[str, Query(description="言葉の表層形")],
+        pronunciation: Annotated[str, Query(description="言葉の発音（カタカナ）")],
+        accent_type: Annotated[
+            int, Query(description="アクセント型（音が下がる場所を指す）")
+        ],
+        word_uuid: Annotated[str, FAPath(description="更新する言葉のUUID")],
+        word_type: Annotated[
+            WordTypes | None,
+            Query(
+                description="PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか"
+            ),
+        ] = None,
+        priority: Annotated[
+            int | None,
+            Query(
+                ge=MIN_PRIORITY,
+                le=MAX_PRIORITY,
+                description="単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨。",
+            ),
+        ] = None,
+    ) -> Response:
         """
         ユーザー辞書に登録されている言葉を更新します。
-
-        Parameters
-        ----------
-        surface : str
-            言葉の表層形
-        pronunciation: str
-            言葉の発音（カタカナ）
-        accent_type: int
-            アクセント型（音が下がる場所を指す）
-        word_uuid: str
-            更新する言葉のUUID
-        word_type: WordTypes, optional
-            PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか
-        priority: int, optional
-            単語の優先度（0から10までの整数）
-            数字が大きいほど優先度が高くなる
-            1から9までの値を指定することを推奨
         """
         try:
             rewrite_word(
@@ -983,20 +1204,26 @@ def generate_app(
         except HTTPException:
             raise
         except ValidationError as e:
-            raise HTTPException(status_code=422, detail="パラメータに誤りがあります。\n" + str(e))
+            raise HTTPException(
+                status_code=422, detail="パラメータに誤りがあります。\n" + str(e)
+            )
         except Exception:
             traceback.print_exc()
-            raise HTTPException(status_code=422, detail="ユーザー辞書の更新に失敗しました。")
+            raise HTTPException(
+                status_code=422, detail="ユーザー辞書の更新に失敗しました。"
+            )
 
-    @app.delete("/user_dict_word/{word_uuid}", status_code=204, tags=["ユーザー辞書"])
-    def delete_user_dict_word(word_uuid: str):
+    @app.delete(
+        "/user_dict_word/{word_uuid}",
+        status_code=204,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def delete_user_dict_word(
+        word_uuid: Annotated[str, FAPath(description="削除する言葉のUUID")]
+    ) -> Response:
         """
         ユーザー辞書に登録されている言葉を削除します。
-
-        Parameters
-        ----------
-        word_uuid: str
-            削除する言葉のUUID
         """
         try:
             delete_word(word_uuid=word_uuid)
@@ -1005,34 +1232,42 @@ def generate_app(
             raise
         except Exception:
             traceback.print_exc()
-            raise HTTPException(status_code=422, detail="ユーザー辞書の更新に失敗しました。")
+            raise HTTPException(
+                status_code=422, detail="ユーザー辞書の更新に失敗しました。"
+            )
 
-    @app.post("/import_user_dict", status_code=204, tags=["ユーザー辞書"])
+    @app.post(
+        "/import_user_dict",
+        status_code=204,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def import_user_dict_words(
-        import_dict_data: Dict[str, UserDictWord], override: bool
-    ):
+        import_dict_data: Annotated[
+            dict[str, UserDictWord],
+            Body(description="インポートするユーザー辞書のデータ"),
+        ],
+        override: Annotated[
+            bool, Query(description="重複したエントリがあった場合、上書きするかどうか")
+        ],
+    ) -> Response:
         """
         他のユーザー辞書をインポートします。
-
-        Parameters
-        ----------
-        import_dict_data: Dict[str, UserDictWord]
-            インポートするユーザー辞書のデータ
-        override: bool
-            重複したエントリがあった場合、上書きするかどうか
         """
         try:
             import_user_dict(dict_data=import_dict_data, override=override)
             return Response(status_code=204)
         except Exception:
             traceback.print_exc()
-            raise HTTPException(status_code=422, detail="ユーザー辞書のインポートに失敗しました。")
+            raise HTTPException(
+                status_code=422, detail="ユーザー辞書のインポートに失敗しました。"
+            )
 
     @app.get("/supported_devices", response_model=SupportedDevicesInfo, tags=["その他"])
     def supported_devices(
-        core_version: Optional[str] = None,
-    ):
-        supported_devices = get_engine(core_version).supported_devices
+        core_version: str | None = None,
+    ) -> Response:
+        supported_devices = get_core(core_version).supported_devices
         if supported_devices is None:
             raise HTTPException(status_code=422, detail="非対応の機能です。")
         return Response(
@@ -1041,13 +1276,45 @@ def generate_app(
         )
 
     @app.get("/engine_manifest", response_model=EngineManifest, tags=["その他"])
-    def engine_manifest():
-        return engine_manifest_loader.load_manifest()
+    def engine_manifest() -> EngineManifest:
+        return engine_manifest_data
 
-    @app.get("/setting", response_class=HTMLResponse, tags=["設定"])
-    def setting_get(request: Request):
-        settings = setting_loader.load_setting_file()
+    @app.post(
+        "/validate_kana",
+        response_model=bool,
+        tags=["その他"],
+        summary="テキストがAquesTalk 風記法に従っているか判定する",
+        responses={
+            400: {
+                "description": "テキストが不正です",
+                "model": ParseKanaBadRequest,
+            }
+        },
+    )
+    def validate_kana(
+        text: Annotated[str, Query(description="判定する対象の文字列")]
+    ) -> bool:
+        """
+        テキストがAquesTalk 風記法に従っているかどうかを判定します。
+        従っていない場合はエラーが返ります。
+        """
+        try:
+            parse_kana(text)
+            return True
+        except ParseKanaError as err:
+            raise HTTPException(
+                status_code=400,
+                detail=ParseKanaBadRequest(err).dict(),
+            )
 
+    @app.get("/setting", response_class=Response, tags=["設定"])
+    def setting_get(request: Request) -> Response:
+        """
+        設定ページを返します。
+        """
+        settings = setting_loader.load()
+
+        brand_name = engine_manifest_data.brand_name
         cors_policy_mode = settings.cors_policy_mode
         allow_origin = settings.allow_origin
 
@@ -1058,64 +1325,97 @@ def generate_app(
             "ui.html",
             {
                 "request": request,
-                "cors_policy_mode": cors_policy_mode,
+                "brand_name": brand_name,
+                "cors_policy_mode": cors_policy_mode.value,
                 "allow_origin": allow_origin,
             },
         )
 
-    @app.post("/setting", response_class=HTMLResponse, tags=["設定"])
+    @app.post(
+        "/setting",
+        response_class=Response,
+        tags=["設定"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def setting_post(
-        request: Request,
-        cors_policy_mode: Optional[str] = Form(None),  # noqa: B008
-        allow_origin: Optional[str] = Form(None),  # noqa: B008
-    ):
+        cors_policy_mode: CorsPolicyMode = Form(),  # noqa
+        allow_origin: str | None = Form(default=None),  # noqa
+    ) -> Response:
+        """
+        設定を更新します。
+        """
         settings = Setting(
             cors_policy_mode=cors_policy_mode,
             allow_origin=allow_origin,
         )
 
         # 更新した設定へ上書き
-        setting_loader.dump_setting_file(settings)
+        setting_loader.save(settings)
 
-        if allow_origin is None:
-            allow_origin = ""
+        return Response(status_code=204)
 
-        return setting_ui_template.TemplateResponse(
-            "ui.html",
-            {
-                "request": request,
-                "cors_policy_mode": cors_policy_mode,
-                "allow_origin": allow_origin,
-            },
+    # BaseLibraryInfo/VvlibManifestモデルはAPIとして表には出ないが、エディタ側で利用したいので、手動で追加する
+    # ref: https://fastapi.tiangolo.com/advanced/extending-openapi/#modify-the-openapi-schema
+    def custom_openapi() -> Any:
+        if app.openapi_schema:
+            return app.openapi_schema
+        openapi_schema = get_openapi(
+            title=app.title,
+            version=app.version,
+            description=app.description,
+            routes=app.routes,
+            tags=app.openapi_tags,
+            servers=app.servers,
+            terms_of_service=app.terms_of_service,
+            contact=app.contact,
+            license_info=app.license_info,
         )
+        openapi_schema["components"]["schemas"][
+            "VvlibManifest"
+        ] = VvlibManifest.schema()
+        # ref_templateを指定しない場合、definitionsを参照してしまうので、手動で指定する
+        base_library_info = BaseLibraryInfo.schema(
+            ref_template="#/components/schemas/{model}"
+        )
+        # definitionsは既存のモデルを重複して定義するため、不要なので削除
+        del base_library_info["definitions"]
+        openapi_schema["components"]["schemas"]["BaseLibraryInfo"] = base_library_info
+        app.openapi_schema = openapi_schema
+        return openapi_schema
+
+    app.openapi = custom_openapi  # type: ignore[method-assign]
 
     return app
 
 
-if __name__ == "__main__":
+def main() -> None:
     multiprocessing.freeze_support()
 
-    output_log_utf8 = os.getenv("VV_OUTPUT_LOG_UTF8", default="")
-    if output_log_utf8 == "1":
+    output_log_utf8 = decide_boolean_from_env("VV_OUTPUT_LOG_UTF8")
+    if output_log_utf8:
         set_output_log_utf8()
-    elif not (output_log_utf8 == "" or output_log_utf8 == "0"):
-        print(
-            "WARNING:  invalid VV_OUTPUT_LOG_UTF8 environment variable value",
-            file=sys.stderr,
-        )
-
-    default_cors_policy_mode = CorsPolicyMode.localapps
 
     parser = argparse.ArgumentParser(description="VOICEVOX のエンジンです。")
+    # Uvicorn でバインドするアドレスを "localhost" にすることで IPv4 (127.0.0.1) と IPv6 ([::1]) の両方でリッスンできます.
+    # これは Uvicorn のドキュメントに記載されていない挙動です; 将来のアップデートにより動作しなくなる可能性があります.
+    # ref: https://github.com/VOICEVOX/voicevox_engine/pull/647#issuecomment-1540204653
     parser.add_argument(
-        "--host", type=str, default="127.0.0.1", help="接続を受け付けるホストアドレスです。"
+        "--host",
+        type=str,
+        default="localhost",
+        help="接続を受け付けるホストアドレスです。",
     )
-    parser.add_argument("--port", type=int, default=50021, help="接続を受け付けるポート番号です。")
     parser.add_argument(
-        "--use_gpu", action="store_true", help="指定するとGPUを使って音声合成するようになります。"
+        "--port", type=int, default=50021, help="接続を受け付けるポート番号です。"
     )
     parser.add_argument(
-        "--voicevox_dir", type=Path, default=None, help="VOICEVOXのディレクトリパスです。"
+        "--use_gpu", action="store_true", help="GPUを使って音声合成するようになります。"
+    )
+    parser.add_argument(
+        "--voicevox_dir",
+        type=Path,
+        default=None,
+        help="VOICEVOXのディレクトリパスです。",
     )
     parser.add_argument(
         "--voicelib_dir",
@@ -1134,12 +1434,12 @@ if __name__ == "__main__":
     parser.add_argument(
         "--enable_mock",
         action="store_true",
-        help="指定するとVOICEVOX COREを使わずモックで音声合成を行います。",
+        help="VOICEVOX COREを使わずモックで音声合成を行います。",
     )
     parser.add_argument(
         "--enable_cancellable_synthesis",
         action="store_true",
-        help="指定すると音声合成を途中でキャンセルできるようになります。",
+        help="音声合成を途中でキャンセルできるようになります。",
     )
     parser.add_argument(
         "--init_processes",
@@ -1148,7 +1448,9 @@ if __name__ == "__main__":
         help="cancellable_synthesis機能の初期化時に生成するプロセス数です。",
     )
     parser.add_argument(
-        "--load_all_models", action="store_true", help="指定すると起動時に全ての音声合成モデルを読み込みます。"
+        "--load_all_models",
+        action="store_true",
+        help="起動時に全ての音声合成モデルを読み込みます。",
     )
 
     # 引数へcpu_num_threadsの指定がなければ、環境変数をロールします。
@@ -1159,8 +1461,8 @@ if __name__ == "__main__":
         type=int,
         default=os.getenv("VV_CPU_NUM_THREADS") or None,
         help=(
-            "音声合成を行うスレッド数です。指定しないと、代わりに環境変数VV_CPU_NUM_THREADSの値が使われます。"
-            "VV_CPU_NUM_THREADSが空文字列でなく数値でもない場合はエラー終了します。"
+            "音声合成を行うスレッド数です。指定しない場合、代わりに環境変数 VV_CPU_NUM_THREADS の値が使われます。"
+            "VV_CPU_NUM_THREADS が空文字列でなく数値でもない場合はエラー終了します。"
         ),
     )
 
@@ -1168,7 +1470,7 @@ if __name__ == "__main__":
         "--output_log_utf8",
         action="store_true",
         help=(
-            "指定するとログ出力をUTF-8でおこないます。指定しないと、代わりに環境変数 VV_OUTPUT_LOG_UTF8 の値が使われます。"
+            "ログ出力をUTF-8でおこないます。指定しない場合、代わりに環境変数 VV_OUTPUT_LOG_UTF8 の値が使われます。"
             "VV_OUTPUT_LOG_UTF8 の値が1の場合はUTF-8で、0または空文字、値がない場合は環境によって自動的に決定されます。"
         ),
     )
@@ -1182,15 +1484,45 @@ if __name__ == "__main__":
             "CORSの許可モード。allまたはlocalappsが指定できます。allはすべてを許可します。"
             "localappsはオリジン間リソース共有ポリシーを、app://.とlocalhost関連に限定します。"
             "その他のオリジンはallow_originオプションで追加できます。デフォルトはlocalapps。"
+            "このオプションは--setting_fileで指定される設定ファイルよりも優先されます。"
         ),
     )
 
     parser.add_argument(
-        "--allow_origin", nargs="*", help="許可するオリジンを指定します。スペースで区切ることで複数指定できます。"
+        "--allow_origin",
+        nargs="*",
+        help=(
+            "許可するオリジンを指定します。スペースで区切ることで複数指定できます。"
+            "このオプションは--setting_fileで指定される設定ファイルよりも優先されます。"
+        ),
     )
 
     parser.add_argument(
-        "--setting_file", type=Path, default=USER_SETTING_PATH, help="設定ファイルを指定できます。"
+        "--setting_file",
+        type=Path,
+        default=USER_SETTING_PATH,
+        help="設定ファイルを指定できます。",
+    )
+
+    parser.add_argument(
+        "--preset_file",
+        type=Path,
+        default=None,
+        help=(
+            "プリセットファイルを指定できます。"
+            "指定がない場合、環境変数 VV_PRESET_FILE、--voicevox_dirのpresets.yaml、"
+            "実行ファイルのディレクトリのpresets.yamlを順に探します。"
+        ),
+    )
+
+    parser.add_argument(
+        "--disable_mutable_api",
+        action="store_true",
+        help=(
+            "辞書登録や設定変更など、エンジンの静的なデータを変更するAPIを無効化します。"
+            "指定しない場合、代わりに環境変数 VV_DISABLE_MUTABLE_API の値が使われます。"
+            "VV_DISABLE_MUTABLE_API の値が1の場合は無効化で、0または空文字、値がない場合は無視されます。"
+        ),
     )
 
     args = parser.parse_args()
@@ -1198,35 +1530,55 @@ if __name__ == "__main__":
     if args.output_log_utf8:
         set_output_log_utf8()
 
-    cpu_num_threads: Optional[int] = args.cpu_num_threads
+    # Synthesis Engine
+    use_gpu: bool = args.use_gpu
+    voicevox_dir: Path | None = args.voicevox_dir
+    voicelib_dirs: list[Path] | None = args.voicelib_dir
+    runtime_dirs: list[Path] | None = args.runtime_dir
+    enable_mock: bool = args.enable_mock
+    cpu_num_threads: int | None = args.cpu_num_threads
+    load_all_models: bool = args.load_all_models
 
-    synthesis_engines = make_synthesis_engines(
-        use_gpu=args.use_gpu,
-        voicelib_dirs=args.voicelib_dir,
-        voicevox_dir=args.voicevox_dir,
-        runtime_dirs=args.runtime_dir,
+    cores = initialize_cores(
+        use_gpu=use_gpu,
+        voicelib_dirs=voicelib_dirs,
+        voicevox_dir=voicevox_dir,
+        runtime_dirs=runtime_dirs,
         cpu_num_threads=cpu_num_threads,
-        enable_mock=args.enable_mock,
-        load_all_models=args.load_all_models,
+        enable_mock=enable_mock,
+        load_all_models=load_all_models,
     )
-    assert len(synthesis_engines) != 0, "音声合成エンジンがありません。"
-    latest_core_version = get_latest_core_version(versions=synthesis_engines.keys())
+    tts_engines = make_tts_engines_from_cores(cores)
+    assert len(tts_engines) != 0, "音声合成エンジンがありません。"
+    latest_core_version = get_latest_core_version(versions=list(tts_engines.keys()))
 
-    cancellable_engine = None
-    if args.enable_cancellable_synthesis:
-        cancellable_engine = CancellableEngine(args)
+    # Cancellable Engine
+    enable_cancellable_synthesis: bool = args.enable_cancellable_synthesis
+    init_processes: int = args.init_processes
 
-    root_dir = args.voicevox_dir if args.voicevox_dir is not None else engine_root()
+    cancellable_engine: CancellableEngine | None = None
+    if enable_cancellable_synthesis:
+        cancellable_engine = CancellableEngine(
+            init_processes=init_processes,
+            use_gpu=use_gpu,
+            voicelib_dirs=voicelib_dirs,
+            voicevox_dir=voicevox_dir,
+            runtime_dirs=runtime_dirs,
+            cpu_num_threads=cpu_num_threads,
+            enable_mock=enable_mock,
+        )
 
-    setting_loader = SettingLoader(args.setting_file)
+    root_dir: Path | None = voicevox_dir
+    if root_dir is None:
+        root_dir = engine_root()
 
-    settings = setting_loader.load_setting_file()
+    setting_loader = SettingHandler(args.setting_file)
 
-    cors_policy_mode = (
-        args.cors_policy_mode
-        if args.cors_policy_mode is not None
-        else settings.cors_policy_mode
-    )
+    settings = setting_loader.load()
+
+    cors_policy_mode: CorsPolicyMode | None = args.cors_policy_mode
+    if cors_policy_mode is None:
+        cors_policy_mode = settings.cors_policy_mode
 
     allow_origin = None
     if args.allow_origin is not None:
@@ -1234,15 +1586,45 @@ if __name__ == "__main__":
     elif settings.allow_origin is not None:
         allow_origin = settings.allow_origin.split(" ")
 
+    # Preset Manager
+    # preset_pathの優先順: 引数、環境変数、voicevox_dir、実行ファイルのディレクトリ
+    # ファイルの存在に関わらず、優先順で最初に指定されたパスをプリセットファイルとして使用する
+    preset_path: Path | None = args.preset_file
+    if preset_path is None:
+        # 引数 --preset_file の指定がない場合
+        env_preset_path = os.getenv("VV_PRESET_FILE")
+        if env_preset_path is not None and len(env_preset_path) != 0:
+            # 環境変数 VV_PRESET_FILE の指定がある場合
+            preset_path = Path(env_preset_path)
+        else:
+            # 環境変数 VV_PRESET_FILE の指定がない場合
+            preset_path = root_dir / "presets.yaml"
+
+    preset_manager = PresetManager(
+        preset_path=preset_path,
+    )
+
+    disable_mutable_api: bool = args.disable_mutable_api | decide_boolean_from_env(
+        "VV_DISABLE_MUTABLE_API"
+    )
+
     uvicorn.run(
         generate_app(
-            synthesis_engines,
+            tts_engines,
+            cores,
             latest_core_version,
             setting_loader,
+            preset_manager=preset_manager,
+            cancellable_engine=cancellable_engine,
             root_dir=root_dir,
             cors_policy_mode=cors_policy_mode,
             allow_origin=allow_origin,
+            disable_mutable_api=disable_mutable_api,
         ),
         host=args.host,
         port=args.port,
     )
+
+
+if __name__ == "__main__":
+    main()
