@@ -1,10 +1,13 @@
 from unittest import TestCase
 
-from voicevox_engine.utility import get_latest_core_version, parse_core_version
+from voicevox_engine.utility.core_version_utility import (
+    get_latest_core_version,
+    parse_core_version,
+)
 
 
 class TestCoreVersion(TestCase):
-    def test_parse_core_version(self):
+    def test_parse_core_version(self) -> None:
         parse_core_version("0.0.0")
         parse_core_version("0.1.0")
         parse_core_version("0.10.0")
@@ -13,7 +16,7 @@ class TestCoreVersion(TestCase):
         parse_core_version("0.14.0-preview.1")
         parse_core_version("0.14.0-preview.10")
 
-    def test_get_latest_core_version(self):
+    def test_get_latest_core_version(self) -> None:
         self.assertEqual(
             get_latest_core_version(
                 versions=[
