@@ -1,16 +1,31 @@
 from enum import Enum
-from typing import List, Optional
+from typing import List, Literal, NewType, Optional
 
 from pydantic import BaseModel, Field
+
+# NOTE: 循環importを防ぐためにとりあえずここに書いている
+# FIXME: 他のmodelに依存せず、全modelから参照できる場所に配置する
+StyleId = NewType("StyleId", int)
+StyleType = Literal["talk", "singing_teacher", "frame_decode", "sing"]
 
 
 class SpeakerStyle(BaseModel):
     """
-    スピーカーのスタイル情報
+    話者のスタイル情報
     """
 
     name: str = Field(title="スタイル名")
-    id: int = Field(title="スタイルID")
+    id: StyleId = Field(title="スタイルID")
+    type: Optional[StyleType] = Field(
+        default="talk",
+        title=(
+            "スタイルの種類。"
+            "talk:音声合成クエリの作成と音声合成が可能。"
+            "singing_teacher:歌唱音声合成用のクエリの作成が可能。"
+            "frame_decode:歌唱音声合成が可能。"
+            "sing:歌唱音声合成用のクエリの作成と歌唱音声合成が可能。"
+        ),
+    )
 
 
 class SpeakerSupportPermittedSynthesisMorphing(str, Enum):
@@ -29,34 +44,35 @@ class SpeakerSupportedFeatures(BaseModel):
     """
 
     permitted_synthesis_morphing: SpeakerSupportPermittedSynthesisMorphing = Field(
-        title="モーフィング機能への対応", default=SpeakerSupportPermittedSynthesisMorphing(None)
+        title="モーフィング機能への対応",
+        default=SpeakerSupportPermittedSynthesisMorphing(None),
     )
 
 
 class CoreSpeaker(BaseModel):
     """
-    コアに含まれるスピーカー情報
+    コアに含まれる話者情報
     """
 
     name: str = Field(title="名前")
-    speaker_uuid: str = Field(title="スピーカーのUUID")
-    styles: List[SpeakerStyle] = Field(title="スピーカースタイルの一覧")
-    version: str = Field("スピーカーのバージョン")
+    speaker_uuid: str = Field(title="話者のUUID")
+    styles: List[SpeakerStyle] = Field(title="スタイルの一覧")
+    version: str = Field("話者のバージョン")
 
 
 class EngineSpeaker(BaseModel):
     """
-    エンジンに含まれるスピーカー情報
+    エンジンに含まれる話者情報
     """
 
     supported_features: SpeakerSupportedFeatures = Field(
-        title="スピーカーの対応機能", default_factory=SpeakerSupportedFeatures
+        title="話者の対応機能", default_factory=SpeakerSupportedFeatures
     )
 
 
 class Speaker(CoreSpeaker, EngineSpeaker):
     """
-    スピーカー情報
+    話者情報
     """
 
     pass
@@ -67,10 +83,14 @@ class StyleInfo(BaseModel):
     スタイルの追加情報
     """
 
-    id: int = Field(title="スタイルID")
+    id: StyleId = Field(title="スタイルID")
     icon: str = Field(title="当該スタイルのアイコンをbase64エンコードしたもの")
-    portrait: Optional[str] = Field(title="当該スタイルのportrait.pngをbase64エンコードしたもの")
-    voice_samples: List[str] = Field(title="voice_sampleのwavファイルをbase64エンコードしたもの")
+    portrait: Optional[str] = Field(
+        title="当該スタイルのportrait.pngをbase64エンコードしたもの"
+    )
+    voice_samples: List[str] = Field(
+        title="voice_sampleのwavファイルをbase64エンコードしたもの"
+    )
 
 
 class SpeakerInfo(BaseModel):
