@@ -3,7 +3,7 @@ import sys
 import traceback
 from pathlib import Path
 
-from appdirs import user_data_dir
+from platformdirs import user_data_dir
 
 
 def engine_root() -> Path:
@@ -33,7 +33,7 @@ def is_development() -> bool:
     return True
 
 
-def get_save_dir():
+def get_save_dir() -> Path:
     # FIXME: ファイル保存場所をエンジン固有のIDが入ったものにする
     # FIXME: Windowsは`voicevox-engine/voicevox-engine`ディレクトリに保存されているので
     # `VOICEVOX/voicevox-engine`に変更する
